fix: read project history synchronously during setup
Previously, noun (and project.nvim) would use the same asynchronous
function to load the project history for the initial read when `setup()`
is called, and when the project history file changes. This lead to a
state where the project history might not be fully loaded by the time
telescope tries to draw the picker window.

This changes it to do the initial load synchronously, so that we can be
sure that the project history is fully loaded by the time the `setup()`
function returns.

Loads that happen while editing due to history file changes will still
be done asynchronously.